requirements.txt
matplotlib==3.6.3
pandas==1.5.3
numpy==1.24.1
scipy==1.10.0
pillow==9.4.0
scikit-learn==1.2.0
opencv==4.7.0
ipykernel==6.20.2
tqdm==4.64.1
openpyxl==3.0.10